[DM-21120] Updates to allow for environment config
First, we should use the same image everywhere, and only override
the URL for the MOTD.  So move index.html to index.html.template.
Instead of a hardcoded URL, use ${MOTD_URL}.

Change entrypoint.sh to create index.html on startup, using the
environment variable MOTD_URL.  This uses envsubst to replace
the value with the URL.

Update the deployment to pass in an environment variable with the
default being the stable MOTD.

For nginx to write the index.html, the files added should be
owned by nginx.  Edit the Dockerfile to make it so.  Make
the Dockerfile use the new entrypoint that does the shell
replacement.

Add helper scripts to build an image and push an image.